requirements.txt
astroid==2.5.6
asyncio==3.4.3
grpcio==1.37.1
grpcio-tools==1.37.1
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
mypy==0.812
mypy-extensions==0.4.3
protobuf==3.16.0
pylint==2.8.2
six==1.16.0
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.10.0.0
wrapt==1.12.1
